Rename Kube to Kubes
Move props to the bottom
